xen/passthrough:coproc: Introduce new APIs and platform callbacks to handle coproc device
It was considered to keep two sets of API+callback:
1. assign_coproc/deassign_coproc
2. disable_coproc/enable_coproc

First set is intended to be called during vcoproc initialization/destruction
for doing all required work needed for the following IOMMU switching in runtime,
etc, therefore it may contain memory allocation, busy loops, etc.
Second set is intended to be called during vcoproc context switch only and
as the result it must be as small as possible and can be called from
the interrupt context.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>