xen/arm: iommu: Panic if not all IOMMUs are initialized
At the moment, the platform can come up with only part of the IOMMUs
initialized. This could lead to a failure later on when building the
hardware domain or even trying to assign a device to a guest.

To avoid unwanted behavior, Xen will not continue if one of the IOMMUs
has not been initialized correctly.

Signed-off-by: Julien Grall <julien.grall@arm.com>

Message-Id: <20190820122255.9864-1-julien.grall@arm.com>